for Backend test/requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
black==24.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
django-database-url==1.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
gunicorn==22.0.0
idna==3.7
mssql-django==1.5
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.11.1
PyJWT==2.9.0
pyodbc==5.1.0
pypng==0.20220715.0
pytz==2024.1
qrcode==7.4.2
requests==2.32.3
sqlparse==0.5.1
twilio==9.2.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
waitress==3.0.0
yarl==1.9.4
